requirements.txt
numpy==1.21.1
opencv-python
tqdm
py360convert
transformers
torch
torchvision
torchaudio
pycocotools
pillow
matplotlib
open3d